requirements.txt
croniter==6.0.0
python_dateutil==2.9.0
PyYAML==6.0.2
Requests==2.32.3